Refactor: エージェント種別ごとに実装ファイルを分割
