+++ src/app/agents/tic-tac-toe-agent.ts
@@ -1,0 +1,5 @@
+import { GameState, Player } from '../game.types';
+
+export interface TicTacToeAgent {
+  pickMove(state: GameState, player: Player): number;
+}
